+++ lib/product_hunt/api.rb
@@ -1,5 +1,3 @@
-# TODO: factor out a base class btw User and Post
-
 module ProductHunt
   module API
     include ProductHunt::API::Posts